Design fixes (See Description)
Moved icons-ie.js out of the tree that manifest requires so it won't load double icons. This will be called conditionally via HTML for IE6 browsers.

Removed bootstrap-popover since it looks like we won't be using it for now.

Added CSS classes to bootstrap overrides so the overlay works for meny irregardless of content availability in the body. This fixes #74

Signed-off-by: Kenneth Lim <kenlimmj@gmail.com>