I THINK I THINK I THAW A TWEEDY BURD! (See Description)
Added a Postmessage library to interface with the YouTube JS API via iframe.

Super ghetto solution for pausing the lectures on slide change. Since it's non-trivial to find out whether we're going forward or going backwards, and we can only poll for the current slide after the slide has been changed, we pause the lecture on the current slide, the slide before, and the slide after. Tadah :D. This also closes #14.

Fixed up bits and pieces of #70. Still WIP

enquireJS more or less working ok unless some tootsie decides to intentionally change the size of his screen all the time. Closes #95

Fixed the problem with the intrinsic ratio and the black borders on top when playing. Specifically, added a 5px border the same color as the background. Ghetto is good :P. Closes #104

Fine-tuned the position of the navigation arrows, and in doing so closes #105. Unfortunately the tooltips look fugly.

Signed-off-by: Kenneth Lim <kenlimmj@gmail.com>